backend/requirements.txt
# Django Core
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.1

# Authentification & Sécurité
djangorestframework-simplejwt>=5.3.1
django-allauth>=0.61.1
argon2-cffi>=23.1.0
django-axes>=6.1.1

# CAPTCHA
requests>=2.31.0

# Base de données
psycopg[binary]>=3.1.18
dj-database-url>=2.1.0

# Redis & Cache
redis>=5.0.1
django-redis>=5.4.0

# Celery (Tâches asynchrones)
celery>=5.3.6
django-celery-beat>=2.6.0
django-celery-results>=2.5.1

# WebSocket (Temps réel)
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.1.0

# API & Validation
drf-spectacular>=0.27.1
django-filter>=23.5
Pillow>=12.1.0

# Utilitaires
python-dotenv>=1.0.1
python-decouple>=3.8

# Monitoring & Error Tracking
sentry-sdk>=1.40.0

# Développement
django-debug-toolbar>=4.2.0
ipython>=8.21.0

# Tests
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=22.5.1

# Qualité de code
flake8>=7.0.0
black>=24.1.1
isort>=5.13.2
